build: change client to use gulp
This commit finishes up the npm scripts -> gulp conversion process.
However, it should be noted that there is one instance when npm
scripts still need to be used: when running a server that needs
to be killed (such as in the e2e tests). Otherwise, gulp is
used and _should be_ used for all tasks.

Note that the way that commands are run is still done via
npm scripts, that is the "API" of the build process is still
exposed via npm. Thus, for example to run the pre-commit scripts
from the terminal you would run `npm run preCommit`. This just
_happens_ to call a gulp task via `npx gulp preCommit`. This way
the build process is more familiar, and users/me don't have to
think about the fact that the tasks are running in gulp
instead of npm.

This is also nice because oh-my-zsh provides a [npm auto-completion
plugin](https://github.com/robbyrussell/oh-my-zsh/wiki/Plugins#np://github.com/robbyrussell/oh-my-zsh/kiwi/Plugins#npm)
but as far as I can tell does _not_ provide a gulp one. So this
makes using tasks much more user-friendly. Furthermore, this
method means that there is a clear distinction between
"public" tasks and "private" internal ones. The public ones
are exported and have an npm script API. The private ones are
_not_ exported and don't have a npm script API.

Another benefit of this approach, is that the npm script API
tasks can have **any** name desired, rather than the severe
restrictions that come from using functions as task using
the modern gulp style. This means that the gulp task back-end
can have function style names (no colons or dashes in the names)
and the npm script task front-end can have standard script names
(with colons and dashes).

+++ gulpfile.babel.ts
@@ -1,9 +1,7 @@
-import { series } from 'gulp';
+import { parallel, series } from 'gulp';
+import isReachable from 'is-reachable';
 import terminalSpawn from 'terminal-spawn';
 import waitOn from 'wait-on';
-
-// monkey-patch parallel to be series
-const parallel = series;
 
 /* *****************************************************************************
  * Private
@@ -34,54 +32,61 @@ const _commands = {
 
 // not a gulp task!
 const _runStorybook = () =>
-  terminalSpawn('npm run storybook', { cwd: _clientDirectory, shell: false });
+  terminalSpawn('npm run storybook:start', {
+    cwd: _clientDirectory,
+  });
 
 // not a gulp task!
 const _serverStart = () =>
   terminalSpawn(_commands.npm.startServerProduction, {
     cwd: _serverDirectory,
-    shell: false,
   });
 
 // not a gulp task!
 const _clientStart = () =>
-  terminalSpawn(_commands.npm.start, { cwd: _clientDirectory, shell: false });
+  terminalSpawn(_commands.npm.start, { cwd: _clientDirectory });
 
 // Tasks
 
 const _cypressStorybookRun = () =>
   terminalSpawn(_commands.npm.cypressStorybookRun, {
     cwd: _clientDirectory,
-    shell: false,
   }).promise;
 
 const _cypressE2eRun = () =>
   terminalSpawn(_commands.npm.cypressE2eRun, {
     cwd: _clientDirectory,
-    shell: false,
   }).promise;
 
-const _waitOnUrl = async (url: string) =>
+const _waitOnUrl = async (url: string, reverse: boolean = false) =>
   waitOn({
+    reverse,
     resources: [url],
   });
+
+const _isUrlUp = async (url: string) => isReachable(url);
 
 // Cypress
 
 const _testStorybook = async () =>
   new Promise(async (resolve, reject) => {
     try {
-      const storybookSpawn = _runStorybook();
-      storybookSpawn.promise.catch(reason => reject(reason));
-      await _waitOnUrl(_storybookUrl);
-      const cypressProcess = await _cypressStorybookRun();
-      storybookSpawn.process.kill();
-      await storybookSpawn.promise;
-
-      if (cypressProcess.status !== 0) {
-        reject(cypressProcess.status);
+      const isStorybookAlreadyRunning = await _isUrlUp(_storybookUrl);
+      if (isStorybookAlreadyRunning) {
+        reject('storybook is already running!!');
       } else {
-        resolve(0);
+        const storybookSpawn = _runStorybook();
+        storybookSpawn.promise.catch(reason => reject(reason));
+        await _waitOnUrl(_storybookUrl);
+        const cypressProcess = await _cypressStorybookRun();
+        storybookSpawn.process.kill();
+        await _waitOnUrl(_storybookUrl, true);
+
+        if (cypressProcess.status !== 0) {
+          reject(cypressProcess.status);
+        } else {
+          resolve(0);
+        }
       }
     } catch (error) {
       reject(error);
@@ -91,23 +96,34 @@ const _testStorybook = async () =>
 const _testEuclidE2e = async () =>
   new Promise(async (resolve, reject) => {
     try {
-      const serverSpawn = _serverStart();
-      serverSpawn.promise.catch(reason => reject(reason));
-      await _waitOnUrl(_serverUrl);
-      const clientSpawn = _clientStart();
-      clientSpawn.promise.catch(reason => reject(reason));
-      await _waitOnUrl(_clientUrl);
-      const cypressProcess = await _cypressE2eRun();
-
-      serverSpawn.process.kill();
-      clientSpawn.process.kill();
-      await serverSpawn.promise;
-      await clientSpawn.promise;
-
-      if (cypressProcess.status !== 0) {
-        reject(cypressProcess.status);
+      const isServerAlreadyRunning = await _isUrlUp(_serverUrl);
+      if (isServerAlreadyRunning) {
+        reject('server is already running!!');
       } else {
-        resolve(cypressProcess.status);
+        const serverSpawn = _serverStart();
+        serverSpawn.promise.catch(reason => reject(reason));
+        await _waitOnUrl(_serverUrl);
+
+        const isClientAlreadyRunning = await _isUrlUp(_clientUrl);
+        if (isClientAlreadyRunning) {
+          reject('client is already running!!');
+        } else {
+          const clientSpawn = _clientStart();
+          clientSpawn.promise.catch(reason => reject(reason));
+          await _waitOnUrl(_clientUrl);
+          const cypressProcess = await _cypressE2eRun();
+
+          serverSpawn.process.kill();
+          clientSpawn.process.kill();
+          await _waitOnUrl(_serverUrl, true);
+          await _waitOnUrl(_clientUrl, true);
+
+          if (cypressProcess.status !== 0) {
+            reject(cypressProcess.status);
+          } else {
+            resolve(cypressProcess.status);
+          }
+        }
       }
     } catch (error) {
       reject(error);
@@ -121,58 +137,42 @@ const _runCypressTests = series(_testStorybook, _testEuclidE2e);
 const _clientInstall = () =>
   terminalSpawn(_commands.npm.install, {
     cwd: _clientDirectory,
-    shell: false,
   }).promise;
 
 const _clientInstallCi = () =>
   terminalSpawn(_commands.npm.ci, {
     cwd: _clientDirectory,
-    shell: false,
   }).promise;
 
 const _clientPreCommit = () =>
   terminalSpawn(_commands.npm.preCommit, {
-    cwd: _serverDirectory,
-    shell: false,
+    cwd: _clientDirectory,
   }).promise;
 
 // Server
 
 const _serverInstall = () =>
-  terminalSpawn(_commands.npm.install, { cwd: _serverDirectory, shell: false })
-    .promise;
+  terminalSpawn(_commands.npm.install, { cwd: _serverDirectory }).promise;
 
 const _serverInstallCi = () =>
-  terminalSpawn(_commands.npm.ci, { cwd: _serverDirectory, shell: false })
-    .promise;
+  terminalSpawn(_commands.npm.ci, { cwd: _serverDirectory }).promise;
 
 const _serverPreCommit = () =>
   terminalSpawn(_commands.npm.preCommit, {
     cwd: _serverDirectory,
-    shell: false,
   }).promise;
 
 // Root
 
-const _npmInstall = () =>
-  terminalSpawn(_commands.npm.install, { shell: false }).promise;
+const _gitStatus = () => terminalSpawn('git status').promise;
 
-const _npmCi = () => terminalSpawn(_commands.npm.ci, { shell: false }).promise;
-
-const _gitStatus = () => terminalSpawn('git status', { shell: false }).promise;
-
-const _sleep = (seconds: number) =>
-  terminalSpawn(`sleep ${seconds}`, { shell: false }).promise;
+const _sleep = (seconds: number) => terminalSpawn(`sleep ${seconds}`).promise;
 
 const _sleepForPreview = () => _sleep(_sleepPreviewSeconds);
 
 // Root Combination Tasks
 
 const _gitReview = series(_gitStatus, _sleepForPreview);
-
-const _postInstallStandard = parallel(_clientInstall, _serverInstall);
-
-const _postinstallCi = parallel(_clientInstallCi, _serverInstallCi);
 
 const _clientServerPreCommit = parallel(_clientPreCommit, _serverPreCommit);
 
@@ -182,15 +182,20 @@ const _clientServerPreCommit = parallel(_clientPreCommit, _serverPreCommit);
 
 // Root
 
-const install = series(_npmInstall, _postInstallStandard);
+const postInstallStandard = parallel(_clientInstall, _serverInstall);
 
-const installCi = series(_npmCi, _postinstallCi);
+const postinstallCi = parallel(_clientInstallCi, _serverInstallCi);
 
 const preCommit = series(
   _gitReview,
   parallel(_clientServerPreCommit, _runCypressTests),
 );
 
-export { install, installCi, preCommit };
+export {
+  postInstallStandard,
+  postinstallCi,
+  preCommit,
+  _clientServerPreCommit,
+};
 
 export default preCommit;
